将ping的功能放到beacon中，减少程序运行时间和节点发送数据包开销

+++ AdHocForwardGUI/MainForm.cs
@@ -74,10 +74,10 @@ namespace AdHocBaseApp
                 for (int j = 0; j < org.Nodes.Count; j++)
                 {
                     pen.DashStyle = System.Drawing.Drawing2D.DashStyle.Solid;
-                    gc.FillEllipse(brush, (float)org.Nodes[j].X - r/2 + offsetX,
+                    gc.FillEllipse(brush, (float)org.Nodes[j].X - r / 2 + (float)offsetX,
                         (float)org.Nodes[j].Y - r/2 + offsetY, r, r);
                     gc.DrawString(org.Nodes[j].Id.ToString(),new Font("arial",10) , brush,
-                        (float)org.Nodes[j].X + offsetX, (float)org.Nodes[j].Y + offsetY);
+                        (float)org.Nodes[j].X + (float)offsetX, (float)org.Nodes[j].Y + (float)offsetY);
 
                         //Utility.Distance(org.Nodes[j - 1].X, org.Nodes[j - 1].Y, org.Nodes[j].X,org.Nodes[j].Y) <= global.nodeMaxDist)
                     foreach (Neighbor nb in new List<Neighbor>(org.Nodes[j].Neighbors.Values))
@@ -85,15 +85,15 @@ namespace AdHocBaseApp
                         Reader node = nb.node;
                         gc.DrawLine(pen, (float)org.Nodes[j].X + offsetX,
                             (float)org.Nodes[j].Y + offsetY,
-                            (float)node.X + offsetX, (float)node.Y + offsetY);
+                            (float)node.X + offsetX, (float)node.Y + (float)offsetY);
                     }
                     pen.DashStyle = System.Drawing.Drawing2D.DashStyle.Dot;
                     foreach (int o in org.Nodes[j].NearbyObjectCache.Keys)
                     {
                         ObjectNode node = global.objects[o];
-                        gc.DrawLine(pen, (float)org.Nodes[j].X + offsetX,
-                            (float)org.Nodes[j].Y + offsetY,
-                            (float)node.X + offsetX, (float)node.Y + offsetY);
+                        gc.DrawLine(pen, (float)org.Nodes[j].X + (float)offsetX,
+                            (float)org.Nodes[j].Y + (float)offsetY,
+                            (float)node.X + (float)offsetX, (float)node.Y + (float)offsetY);
                     }
                 }
             }
